Use the dispatch table to call clGetPlatformInfo
Currently, we call clGetPlatformInfo by directly loading the function from the
vendor library. Although this is suggested by the ICD spec, many vendors do not
directly expose clGetPlatformInfo, which breaks ICD loading on such libraries.

Fix this by calling clGetPlatformInfo using the dispatch table instead, which is
guaranteed to work, since clGetPlatformInfo is required to be in the table. This
is a looser interpretation of the spec, but it unbreaks many vendor libraries.